refactor: Update Preview and Section_1 components in new_company step_1
Update the Preview component in the new_company step_1 to fix a bug where the innerStep state was being decremented by 2 instead of 1 when the "Add New Company" button was clicked. This ensures that the innerStep state is updated correctly.

Update the Section_1 component in the new_company step_1 to add a new Select component for selecting a country. This allows the user to choose a country from a dropdown menu. The selected country is stored in the state and can be accessed by the handleSelectSector function.

These changes improve the functionality and user experience of the new_company step_1 page.